fix a bug
Why are the states in stack messed up?
When we push a Elenment(named before), function Push in Stack.h recieves a conference, which is not a pointer. As a result, when process come to an end, the Elenment we create is released, causing the execution of Element's destructor.
Solution:
Let stack's element to be pointer.